+++ assets/src/components/editing/editor/normalizers/block.ts
@@ -15,10 +15,12 @@ export const normalize = (editor: Editor, node: ModelElement | FormattedText, pa
         // so deletion removes the inner block and causes validation errors
         if (node.type === 'p' && parent.type === 'code') {
           Transforms.removeNodes(editor, { at: parentPath });
+          console.warn('Normalizing content: Special case code, removing node', node.type);
           return;
         }
 
         Transforms.removeNodes(editor, { at: path });
+        console.warn('Normalizing content: Removing non block node', node.type);
         return;
       }
     }
@@ -27,6 +29,7 @@ export const normalize = (editor: Editor, node: ModelElement | FormattedText, pa
     if (Editor.isBlock(editor, node) && !schema[node.type].isTopLevel) {
       if (Editor.isEditor(parent)) {
         Transforms.unwrapNodes(editor, { at: path });
+        console.warn('Normalizing content: Unwrapping top level block node', node.type);
         return;
       }
     }