+++ CHANGELOG.md
@@ -6,7 +6,28 @@
 
 ### Enhancements
 
+- Add tab for authoring 'explanation' feedback to all activities
+
+## 0.21.5 (2022-09-01)
+
+### Bug Fixes
+
+- Fix an issue where activity content containing backslashes prior to parser escaping breaks evaluation
+- Fix an issue where survey reset wasn't working properly
+
+## 0.21.4 (2022-08-23)
+
+### Bug Fixes
+
+- Fix an issue where content cant be removed if there is only one top-level element
+
+### Enhancements
+
 - Add support for inline 'term' markup
+- Add support for branching activities and page automation
+- Improve toolbar styles and usability
+- Add ability to browse all pages in a course project
+- Add Math input question type
 
 ## 0.21.3 (2022-08-05)